santoni: Start camera provider after /data decryption
class hal starts before data decryption, which may cause issues when
the camera HAL attempts to write to /data. Furthermore, at this period,
the camera daemon will not be up, and we end up in the retry loop in
attempt to probe the sensors, which fails anyway unless the camera
daemon is up.

I believe the reason why some people with encryption have broken
cameras was due to the fact that the sensor probe retries maxed out by
the time decryption was completed. Reducing the time frame needed to
bring up the camera daemon or probe the sensors after the start of the
camera HAL should fix it.

Signed-off-by: Omansh Krishn <omansh11597@gmail.com>